Add tests using WireMock for Weather API

+++ src/main/java/pl/foodflow/api/dto/WeatherDataDTO.java
@@ -16,7 +16,7 @@ public class WeatherDataDTO {
     private String country;
     private String localtime;
 
-    private double temp_c;
+    private Integer temp_c;
     private double temp_f;
     private double wind_mph;
     private double wind_kph;
@@ -26,10 +26,10 @@ public class WeatherDataDTO {
     private double gust_mph;
     private double gust_kph;
 
-    private double precip_mm;
+    private Integer precip_mm;
     private int humidity;
-    private double vis_km;
-    private double uv;
+    private Integer vis_km;
+    private Integer uv;
     private String text;
     private String icon;
 }
